finite-difference-gradient.hh: start revamping finite-difference computation.
The computation of the Jacobian for finite-difference functions will now
be done column-wise (i.e. argument per argument), rather than row-wise
which was particularly slow and inefficient.

Basically, this removes a for loop that led to redundant computation.

For now, only the "simple" rule has been done, and using the
finite-differences should be *much* faster with this rule after this
update.

The 5-point rule still need to be updated in the same manner, but this
is probably going to take some more work.